requirements.txt
Django==5.2.5
django-ipware==5.0.0
gspread==6.0.1
google-auth==2.25.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.1
django-recaptcha==4.0.0
Pillow==10.3.0
whitenoise==6.6.0
django-cors-headers==4.3.1
django-allauth==0.57.0
django-crispy-forms==2.1
requests==2.31.0
python-dotenv==1.0.0
django-extensions==3.2.3
psycopg2-binary==2.9.9
django-security==0.13.0